Pruning: remove unnecessary backward hook used for masking gradients (#63)
After commit f396c34a362731c765370d368877c2ca367ad651, we
Now always apply a pruning mask at the end of mini-batches (this is
because we understand that weights may be updated by SGD+momentum,
even for weights that are masked. Therefore, there is no need for masking
the gradients: we always mask the weights at the end of the mini-batch.

See issue #53 for more details.